ax_enums.mojom 1
Explicitly casting all uses of enums from ax_enums.idl. This is
preparatory work so we can convert these enums to mojo in a follow-up.

Bug: 650275

+++ ash/system/ime_menu/ime_list_view.cc
@@ -123,7 +123,8 @@ class ImeListItemView : public ActionableView {
     node_data->role = ui::AX_ROLE_CHECK_BOX;
     const ui::AXCheckedState checked_state =
         selected_ ? ui::AX_CHECKED_STATE_TRUE : ui::AX_CHECKED_STATE_FALSE;
-    node_data->AddIntAttribute(ui::AX_ATTR_CHECKED_STATE, checked_state);
+    node_data->AddIntAttribute(ui::AX_ATTR_CHECKED_STATE,
+                               static_cast<int32_t>(checked_state));
   }
 
  private:
